requirement.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
requests==2.24.0
urllib3==1.25.9
boto3==1.14.27
line-bot-sdk==1.16.0
bs4==0.0.1